LayoutTests/imported/w3c/web-platform-tests/html/browsers/history/the-location-interface/scripted_click_location_assign_during_load-expected.txt

The popup blocker must be disabled for this test


FAIL location.assign with click during load assert_equals: expected 2 but got 1
